% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATSVD.R
\name{get.rank}
\alias{get.rank}
\title{Estimate Tucker rank of a given tensor}
\usage{
get.rank(Y, sigma, sparse_mode)
}
\arguments{
\item{Y}{Input tensor, with dimension p1*p2*...*pd.}

\item{sigma}{(Estimated) Gaussian noise level}

\item{sparse_mode}{A d-dimensional logicical vector that indicates whether each mode is sparse.}
}
\value{
Estimated rank as a d-dimensinal integer tuple
}
\description{
Estimate Tucker rank of a given tensor
}
