% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATSVD.R
\name{get.sigma}
\alias{get.sigma}
\title{Estimate Gaussian noise level via MAD.}
\usage{
get.sigma(Y)
}
\arguments{
\item{Y}{Input tensor, with dimension p1*p2*...*pd.}
}
\value{
Estimate of standard deviation of Gaussian noise.
}
\description{
Estimate Gaussian noise level via MAD.
}
