% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATSVD.R
\name{HOOI}
\alias{HOOI}
\title{Sparse low-rank tensor decomposition using High-order orthogonal iteration}
\usage{
HOOI(Y, r, sparse_mode, tmax, vartol)
}
\arguments{
\item{Y}{Input tensor, with dimension p1*p2*...*pd.}

\item{r}{(Estimated) Tucker ranks.}

\item{sparse_mode}{A d-dimensional logicical vector that indicates whether each mode is sparse.}

\item{tmax}{Maximal iteration times.}

\item{vartol}{Covergence telorance.}
}
\value{
Estimated subspace in each mode.
}
\description{
Sparse low-rank tensor decomposition using High-order orthogonal iteration
}
