% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATSVD.R
\name{HOSVD}
\alias{HOSVD}
\title{Sparse low-rank tensor decomposition using High-order SVD}
\usage{
HOSVD(Y, r, sparse_mode)
}
\arguments{
\item{Y}{Input tensor, with dimension p1*p2*...*pd.}

\item{r}{(Estimated) Tucker ranks.}

\item{sparse_mode}{A d-dimensional logicical vector that indicates whether each mode is sparse.}
}
\value{
Estimated subspace in each mode.
}
\description{
Sparse low-rank tensor decomposition using High-order SVD
}
