% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATSVD.R
\name{sine.theta}
\alias{sine.theta}
\title{Calculate the sine-theta distance between two subspaces}
\usage{
sine.theta(U, hatU, q)
}
\arguments{
\item{U}{First subspace}

\item{hatU}{Second subspace}

\item{q}{Hilbert-Schmidt norm}
}
\value{
q-sine-theta distances
}
\description{
Calculate the sine-theta distance between two subspaces
}
