\name{STATSVD-package}
\alias{STATSVD-package}
\alias{STATSVD}
\docType{package}
\title{
\packageTitle{STATSVD}
}
\description{
\packageDescription{STATSVD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{STATSVD}
\packageIndices{STATSVD}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{STATSVD}

Maintainer: \packageMaintainer{STATSVD}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
