% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATSVD.R
\name{STATSVD}
\alias{STATSVD}
\title{Sparse low-rank tensor decomposition using STATSVD.
See Zhang and Han (2019) for reference.}
\usage{
STATSVD(Y, r, sigma, tmax, sparse_mode, vartol)
}
\arguments{
\item{Y}{Input tensor, with dimension p1*p2*...*pd.}

\item{r}{(Estimated) Tucker ranks.}

\item{sigma}{(Estimated) noise level.}

\item{tmax}{Maximal iteration times.}

\item{sparse_mode}{A d-dimensional logicical vector that indicates whether each mode is sparse.}

\item{vartol}{Covergence telorance.}
}
\value{
Estimated subspace in each mode.
}
\description{
Sparse low-rank tensor decomposition using STATSVD.
See Zhang and Han (2019) for reference.
}
