% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATSVD.R
\name{vector_sum_square}
\alias{vector_sum_square}
\title{Calculate the Square of l2 norm of a vector.}
\usage{
vector_sum_square(x)
}
\arguments{
\item{x}{Input vector.}
}
\value{
Square of l2 norm of a vector.
}
\description{
Calculate the Square of l2 norm of a vector.
}
